[frontend] consolidated to 2 files
